# doc/api.doc.json
{"openapi":"3.0.0","paths":{"/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"tags":["Auth"]}},"/refresh":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"tags":["Auth"]}},"/contatos":{"get":{"operationId":"ContatoController_get","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Contatos"]},"post":{"operationId":"ContatoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateContatoDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"tags":["Contatos"]}},"/contatos/{id}/account":{"get":{"operationId":"ContatoController_getAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Contatos"]},"post":{"operationId":"ContatoController_updateAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccountDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"tags":["Contatos"]},"delete":{"operationId":"ContatoController_deleteAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"tags":["Contatos"]}},"/contatos/{id}":{"get":{"operationId":"ContatoController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Contatos"]},"put":{"operationId":"ContatoController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateContatoDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"tags":["Contatos"]}},"/contatos/{id}/account/recover":{"patch":{"operationId":"ContatoController_recoverAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"tags":["Contatos"]}},"/produtos":{"post":{"operationId":"ProdutoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProdutoDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Produtos"]},"get":{"operationId":"ProdutoController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Produtos"]}},"/produtos/{id}":{"get":{"operationId":"ProdutoController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Produtos"]},"put":{"operationId":"ProdutoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProdutoDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Produtos"]},"delete":{"operationId":"ProdutoController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Produtos"]}},"/produtos/{id}/historico-valores":{"get":{"operationId":"ProdutoController_getHistoricoValores","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Produtos"]}},"/produtos/{id}/historico-valores/esp-odont":{"get":{"operationId":"ProdutoController_getHistoricoValoresEspOdont","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Produtos"]}},"/produtos/{id}/historico-valores/dentista":{"get":{"operationId":"ProdutoController_getHistoricoValoresDentista","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Produtos"]}}},"info":{"title":"Laboratório Ferreira API","description":"API RESTful do sistema Laboratório Ferreira.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"senha":{"type":"string"}},"required":["senha"]},"ResultDto":{"type":"object","properties":{"sucesso":{"type":"boolean"},"mensagem":{"type":"string"},"dados":{"type":"object"},"erro":{"type":"object"}},"required":["sucesso"]},"CreateAccountDto":{"type":"object","properties":{"username":{"type":"string","example":"nomeusuario"},"email":{"type":"string","example":"exemplo@email.com"},"senha":{"type":"string","minLength":4,"maxLength":20}},"required":["email","senha"]},"CreateContatoDto":{"type":"object","properties":{"nome":{"type":"string","example":"Danillo Silva"},"telefone":{"type":"string","minLength":11,"maxLength":11,"example":"62123456789"},"categorias":{"type":"array","items":{"type":"object"}},"usuario":{"$ref":"#/components/schemas/CreateAccountDto"}},"required":["nome","categorias"]},"UpdateContatoDto":{"type":"object","properties":{"nome":{"type":"string","example":"Danillo Silva"},"telefone":{"type":"string","minLength":11,"maxLength":11,"example":"62123456789"},"categorias":{"type":"array","items":{"type":"object"}}}},"UpdateAccountDto":{"type":"object","properties":{"username":{"type":"string","example":"nomeusuario"},"email":{"type":"string","example":"exemplo@email.com"}}},"CreateProdutoDto":{"type":"object","properties":{"nome":{"type":"string"},"marca":{"type":"string"},"tipo":{"type":"string"},"descricao":{"type":"string"},"valorEspOdont":{"type":"number"},"valorDentista":{"type":"number"}},"required":["nome","tipo","valorEspOdont","valorDentista"]},"UpdateProdutoDto":{"type":"object","properties":{"ativo":{"type":"boolean"},"nome":{"type":"string"},"marca":{"type":"string"},"tipo":{"type":"string"},"descricao":{"type":"string"},"valorEspOdont":{"type":"number"},"valorDentista":{"type":"number"}}}}}}