# src/modules/sistema/auth/dto/login.dto.ts
import { Transform } from 'class-transformer';
import { IsEmail, IsNotEmpty, IsOptional } from 'class-validator';

export class LoginDto {
  @IsOptional()
  @Transform(({ value }) => value.toLowerCase())
  username?: string;

  @IsOptional()
  @IsEmail()
  @Transform(({ value }) => value.toLowerCase())
  email?: string;

  @IsNotEmpty()
  senha: string;
}